AI_server/requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
boto3==1.34.91
botocore==1.34.91
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
datasets==2.19.0
DateTime==5.5
decorator==4.4.2
dill==0.3.8
evaluate==0.4.1
fastapi==0.110.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
moviepy==2.0.0.dev2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mysql==0.0.3
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==5.26.1
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
responses==0.18.0
s3transfer==0.10.1
safetensors==0.4.3
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
utils==1.0.2
uvicorn==0.29.0
xxhash==3.4.1
yarl==1.9.4
zope.interface==6.3
